docs: simplify provider authoring workflow
