+++ Tests/CodexBarTests/ProviderArchitectureGatekeeperTests.swift
@@ -1,8 +1,43 @@
+import AppKit
+import Foundation
 import Testing
+@testable import CodexBar
 @testable import CodexBarCore
 @testable import CodexBarWidget
 
+@MainActor
 struct ProviderArchitectureGatekeeperTests {
+    @Test
+    func `every provider has descriptor and implementation manifest entries`() {
+        let expected = Set(UsageProvider.allCases)
+        let descriptors = Set(ProviderDescriptorRegistry.all.map(\.id))
+        let implementations = Set(ProviderImplementationRegistry.all.map(\.id))
+        let missingDescriptors = expected.subtracting(descriptors).map(\.rawValue).sorted()
+        let missingImplementations = expected.subtracting(implementations).map(\.rawValue).sorted()
+
+        #expect(
+            missingDescriptors.isEmpty,
+            "Missing descriptor manifest entries: \(missingDescriptors.joined(separator: ", "))")
+        #expect(
+            missingImplementations.isEmpty,
+            "Missing implementation manifest entries: \(missingImplementations.joined(separator: ", "))")
+    }
+
+    @Test
+    func `every provider descriptor has a loadable SVG resource`() throws {
+        let resources = try Self.repoRoot()
+            .appending(path: "Sources/CodexBar/Resources", directoryHint: .isDirectory)
+
+        for descriptor in ProviderDescriptorRegistry.all {
+            let resourceName = descriptor.branding.iconResourceName
+            let url = resources.appending(path: "\(resourceName).svg")
+            #expect(
+                FileManager.default.fileExists(atPath: url.path(percentEncoded: false)),
+                "Missing SVG for \(descriptor.id.rawValue): \(resourceName).svg")
+            #expect(NSImage(contentsOf: url) != nil, "Could not load \(resourceName).svg as NSImage")
+        }
+    }
+
     @Test
     func `widget provider choices match selectable descriptor metadata`() {
         let selectable = Set(ProviderDescriptorRegistry.all.filter(\.metadata.widgetSelectable).map(\.id))
@@ -38,5 +73,18 @@ struct ProviderArchitectureGatekeeperTests {
                 descriptor.metadata.shortDisplayName == expected,
                 "Unexpected widget short label for \(descriptor.id.rawValue).")
         }
+    }
+
+    private static func repoRoot() throws -> URL {
+        var directory = URL(filePath: #filePath).deletingLastPathComponent()
+        for _ in 0..<12 {
+            if FileManager.default.fileExists(
+                atPath: directory.appending(path: "Package.swift").path(percentEncoded: false))
+            {
+                return directory
+            }
+            directory.deleteLastPathComponent()
+        }
+        throw CocoaError(.fileNoSuchFile)
     }
 }
