+++ docs/DEVELOPMENT.md
@@ -107,7 +107,7 @@ CodexBar/
 ## Common Tasks
 
 ### Add a New Provider
-See the canonical [provider authoring guide](provider.md#adding-a-new-provider-current-flow) for the complete flow.
+See the canonical [provider authoring guide](provider.md#adding-a-new-provider) for the complete flow.
 
 1. Add the provider identity to `Sources/CodexBarCore/Providers/Providers.swift`.
 2. Add the descriptor and the fetcher, parser, settings-reader, or status-probe pieces the provider needs under
